% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.json-package.R
\docType{package}
\name{dataset.json-package}
\alias{dataset.json}
\alias{dataset.json-package}
\title{dataset.json: Read and Write Dataset-JSON}
\description{
Read and write CDISC Dataset-JSON files (\url{https://www.cdisc.org/dataset-json}).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://rmagno.eu/dataset.json/}
  \item Report bugs at \url{https://github.com/ramiromagno/dataset.json/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
