% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_json.R
\name{ds_json}
\alias{ds_json}
\title{An S7 class to represent Dataset-JSON data}
\usage{
ds_json(
  creationDateTime = class_missing,
  datasetJSONVersion = class_missing,
  fileOID = class_missing,
  asOfDateTime = class_missing,
  originator = class_missing,
  sourceSystem = class_missing,
  sourceSystemVersion = class_missing,
  clinicalData = class_missing,
  referenceData = class_missing
)
}
\arguments{
\item{creationDateTime}{Time of creation of the file containing the document.}

\item{datasetJSONVersion}{Version of Dataset-JSON standard.}

\item{fileOID}{A unique identifier for this file.}

\item{asOfDateTime}{The date/time at which the source database was queried in order to create this document.}

\item{originator}{The organization that generated the Dataset-JSON file.}

\item{sourceSystem}{The computer system or database management system that is the source of the information in this file.}

\item{sourceSystemVersion}{The version of the "SourceSystem" above.}

\item{clinicalData}{Contains clinical data across multiple subjects.}

\item{referenceData}{Contains non-subject data domains.}
}
\description{
A \code{ds_json} classed object is the R counterpart of a Dataset-JSON file.
}
