% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_json.R
\name{ds_json_data}
\alias{ds_json_data}
\title{An S7 class to represent ClinicalData and ReferenceData elements}
\usage{
ds_json_data(
  studyOID = class_missing,
  metaDataVersionOID = class_missing,
  metaDataRef = class_missing,
  itemGroupData = class_missing
)
}
\arguments{
\item{studyOID}{See ODM definition for study OID (ODM/Study/@OID).}

\item{metaDataVersionOID}{See ODM definition for metadata version OID (ODM/Study/MetaDataVersion/@OID).}

\item{metaDataRef}{URL for a metadata file the describing the data.}

\item{itemGroupData}{Object containing dataset information.}
}
\description{
A \code{ds_json_data} classed object is used to represent either ClinicalData and
ReferenceData elements from the Dataset-JSON standard.
}
