% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_json_example.R
\name{ds_json_example_path}
\alias{ds_json_example_path}
\title{Dataset-JSON example file path}
\usage{
ds_json_example_path(name = NULL)
}
\arguments{
\item{name}{The name of Dataset-JSON example. Run \code{\link[=ds_json_example_path]{ds_json_example_path()}} to
check available options, e.g. \code{ds_json_example_path("ae")}.}
}
\value{
A character vector of absolute paths to Dataset-JSON example files.
}
\description{
\code{\link[=ds_json_example_path]{ds_json_example_path()}} retrieves the local path to a bundled Dataset-JSON
example file.
}
\examples{
ds_json_example_path("ae")

ds_json_example_path("adsl")

}
