% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_json.R
\name{ds_json_group_data}
\alias{ds_json_group_data}
\title{An S7 class to represent the ItemGroupData element}
\usage{
ds_json_group_data(
  itemGroupOID = class_missing,
  records = class_missing,
  name = class_missing,
  label = class_missing,
  items = class_missing,
  itemData = class_missing
)
}
\arguments{
\item{itemGroupOID}{See ODM definition for study OID (ODM/Study/@OID).}

\item{records}{The total number of records in a dataset.}

\item{name}{Dataset name.}

\item{label}{Dataset description.}

\item{items}{Basic information about variables (data frame).}

\item{itemData}{Dataset data (data frame).}
}
\description{
A \code{ds_json_group_data} classed object is used to represent ItemGroupData
element from the Dataset-JSON standard.
}
