% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{import_ds_json}
\alias{import_ds_json}
\title{Import a Dataset-JSON file as a list}
\usage{
import_ds_json(file)
}
\arguments{
\item{file}{A path to a Dataset-JSON file.}
}
\description{
\code{\link[=import_ds_json]{import_ds_json()}} is a simple wrapper around
\link[yyjsonr:read_json_file]{read_json_file()} with sensible defaults.
}
\keyword{internal}
