% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ds_json}
\alias{read_ds_json}
\title{Read a Dataset-JSON file}
\usage{
read_ds_json(file)
}
\arguments{
\item{file}{A path to a Dataset-JSON file.}
}
\value{
A \link[=ds_json]{ds_json} object.
}
\description{
\code{\link[=read_ds_json]{read_ds_json()}} reads a Dataset-JSON into R.
}
\examples{
read_ds_json(ds_json_example_path("ae"))

read_ds_json(ds_json_example_path("adsl"))

}
